engine/requirements.txt
aniso8601==8.0.0
APScheduler==3.8.1
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
Click==7.0
configobj==5.0.6
dataclasses==0.8
dbus-python==1.2.4
DBUtils==1.3
decorator==4.2.1
Flask==1.1.1
Flask-APScheduler==1.12.2
Flask-Cors==3.0.8
Flask-Docs==0.5.3
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
google-api-core==2.2.2
google-api-python-client==2.15.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==2.28.1
google-cloud-core==2.1.0
google-cloud-datacatalog==3.4.3
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
gpg==1.13.1
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
ldap3==2.9
libcst==0.3.23
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nftables==0.1
packaging==21.3
perf==0.1
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pygobject==3.28.3
PyJWT==1.4.2
PyMySQL==0.9.3
pyparsing==3.0.6
python-dateutil==2.6.1
python-linux-procfs==0.6.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyudev==0.21.0
PyYAML==6.0
requests==2.26.0
rpm==4.14.3
rsa==4.8
selinux==2.9
sepolicy==1.1
setools==4.3.0
six==1.16.0
slip==0.6.4
slip.dbus==0.6.4
typing-extensions==4.0.0
typing-inspect==0.7.1
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==2.0.2
xmltodict==0.12.0
zipp==3.6.0
